Fix merge conflicts with iso branch
