+++ Assets/Scripts/Tiles/ResourceTile.cs
@@ -1,14 +1,24 @@
-﻿namespace Untitled
+﻿using Untitled.Resource;
+
+namespace Untitled
 {
     namespace Tiles
     {
-        public class ResourceTile
+        public class ResourceTile : IResourceStorage
         {
-            public float Value { get; set; }
-            
-            public ResourceTile(float value)
+            private float val;
+			
+            public void AddResources(ResourceType type, float count) {
+              val += count;
+            }
+
+            public float GetResourceCount(ResourceType type) {
+              return val;
+            }
+
+            public ResourceTile(float startingVal)
             {
-                Value = value;
+                val = startingVal;
             }
         }
     }
